REFACTOR: move census loader function
No actual changes, except for re-ordering some imports
