+++ products/edde/aggregate/PUMS/pums_2000_economics.py
@@ -3,14 +3,15 @@ in an xlsx spreadsheet (Educational attainment data points) cleans them and outp
 that they can be collated using the established collate process"""
 
 import pandas as pd
+
+from aggregate.load_aggregated import load_2000_census_pums_all_data
+from aggregate.aggregation_helpers import order_aggregated_columns
 from utils.PUMA_helpers import clean_PUMAs, dcp_pop_races
 from utils.dcp_population_excel_helpers import (
     race_suffix_mapper,
     map_stat_suffix,
-    load_2000_census_pums_all_data,
 )
 from internal_review.set_internal_review_file import set_internal_review_files
-from aggregate.aggregation_helpers import order_aggregated_columns
 
 # from aggregate.aggregation_helpers import order_aggregated_columns, get_category
