+++ products/edde/aggregate/housing_security/pums_2000_hsq_housing_tenure.py
@@ -1,12 +1,13 @@
 import pandas as pd
+
+from aggregate.load_aggregated import load_2000_census_pums_all_data
+from aggregate.aggregation_helpers import order_aggregated_columns
 from utils.dcp_population_excel_helpers import (
     race_suffix_mapper,
     map_stat_suffix,
-    load_2000_census_pums_all_data,
 )
 from utils.PUMA_helpers import clean_PUMAs, dcp_pop_races
 from internal_review.set_internal_review_file import set_internal_review_files
-from aggregate.aggregation_helpers import order_aggregated_columns
 
 _housing_tenure_name_mapper = {
     "occhu": "units_occupied_2000",
